if user has logged in before admin cannot delete user

+++ admin/src/controllers/userController.ts
@@ -5,16 +5,29 @@ const prisma = new PrismaClient();
 
 const deleteUser = async (req: express.Request, res: express.Response) => {
   const { userId } = req.params;
-
   try {
-    // Delete the user from the database
+    const user = await prisma.user.findUnique({
+      where: {
+        id: parseInt(userId),
+      },
+      select: {
+        firstLoginInAt: true,
+      },
+    });
+    if (!user) {
+      return res.status(404).send({ error: "User not found" });
+    }
+    if (user.firstLoginInAt) {
+      return res
+        .status(202)
+        .send("User has already logged in and cannot be deleted");
+    }
     await prisma.user.delete({
       where: {
         id: parseInt(userId),
       },
     });
-
-    res.status(200).json({ message: "User deleted successfully." });
+    res.status(200).send({ message: "User deleted successfully." });
   } catch (error) {
     console.error("Error deleting user:", error);
     res
